refactor/#155: 패키지명 변경 (components -> component) 및 안쓰는 컴포넌트 제거
