--- feature/history/src/main/java/com/yapp/feature/history/attend/components/SessionAttendanceItem.kt
@@ -1,11 +1,0 @@
-package com.yapp.feature.history.attend.components
-
-import androidx.compose.runtime.Composable
-
-@Composable
-internal fun SessionAttendanceItem(
-    date: String,
-    time: String? = null,
-) {
-
-}
